Revert "add RolloverSupporter"
This reverts commit 07003776
